Move material from root into map folder
