+++ include/radix/component/MeshDrawable.hpp
@@ -2,7 +2,7 @@
 #define RADIX_COMPONENT_MESHDRAWABLE_HPP
 
 #include <radix/component/Component.hpp>
-#include <radix/material/Material.hpp>
+#include <radix/map/material/Material.hpp>
 #include <radix/model/Mesh.hpp>
 
 namespace radix {
