+++ source/map/XmlMapLoader.cpp
@@ -15,7 +15,7 @@
 
 #include <radix/World.hpp>
 #include <radix/model/MeshLoader.hpp>
-#include <radix/material/MaterialLoader.hpp>
+#include <radix/map/material/MaterialLoader.hpp>
 
 using namespace std;
 using namespace tinyxml2;
