+++ source/renderer/TextRenderer.cpp
@@ -3,7 +3,7 @@
 #include <radix/renderer/Renderer.hpp>
 #include <radix/text/FontLoader.hpp>
 #include <radix/shader/ShaderLoader.hpp>
-#include <radix/material/MaterialLoader.hpp>
+#include <radix/map/material/MaterialLoader.hpp>
 #include <radix/core/math/Vector2f.hpp>
 #include <radix/core/math/Vector4f.hpp>
